feat(restaurant_view): add get_restaurant_couriers action to retrieve couriers for a restaurant feat(delivery_calculation): change order_price and distance to float for improved calculation accuracy feat(urls): add new endpoint for retrieving restaurant couriers

+++ menu_app/view/delivery_calculation.py
@@ -17,8 +17,8 @@ class DeliveryCalculationView(APIView):
             )
 
         try:
-            order_price = int(order_price)
-            distance = int(distance) if distance else 0
+            order_price = float(order_price)
+            distance = float(distance) if distance else 0
         except ValueError:
             return Response(
                 {"message": "Некорректный тип параметров", "code": 4},
